+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Fix
+- @javierfh03 - Check if crontab file is empty.
+
 ## [0.1.0] - 2024-02-19
 ## Added
 - @javierfh03 - Develop get cron jobs in crontab manager with unit tests.
@@ -10,4 +14,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - @javierfh03 - Develop modify cron jobs in crontab manager with unit tests.
 - @javierfh03 - Develop delete cron jobs in crontab manager with unit tests.
 
+[Unreleased]: https://github.com/Lagatrix/crontab_lib.git
 [0.1.0]: https://github.com/Lagatrix/crontab_lib.git/releases/tag/0.1.0